Fix fonts a little more.
It turned out that on windows the fonts looked good, but on mac they looked
terrible.  I'm pretty sure that the problem had something to do with the way my openGL driver on windows handled texture mapping through display lists.

It looked like I accidentally got what I wanted on windows, but the Mac interpreted my code correctly (that is, what you would expect), but not what I wanted

So I got rid of the display lists and calculated a font scale based on the window height, l2norm, and vfov and the font remains a constant size throughout its duration.  It can still be adjusted

git-svn-id: https://gforge.sci.utah.edu/svn/map3d/trunk@141 76038eab-591d-0410-bf0b-a17c1341c730