Rearrange a LOT of code on how we handle matlab files
Previously, when we used the FileDialog,
we would load the matlab file for EACH of these:
* to see how many time series
* to see the time series labels
* to see the number of frames
* to load the RMS data curve
* to load the actual data

Now, we read the matlab file once, and pull out the matlabarray for
multiple uses.  We delete that when we close the file.  The also speeds up
the loading of these files.

Alternately, command line usage would load the file once, but made the
file loading a little more sound so we don't keep around the matlab file
info along with the matlabarray.

Windows will also now use the /LARGEADDRESSAWARE linker flag to let it
go up to 3GB per process in a 32bit process

In both ways, map3d will use about 1GB to read the file that produced the
bug:
about 250 MB to open the file
about 450 MB to uncompress the data
about 450 MB to copy the data out
(at which point we can delete the file and the uncompress)

Next Step is to test this out a bit more to make sure I didn't cause any
crazy memory problems, and to get an actual win64 build.