Reorganize P, S, and C options
Before, S and P were independent, resulting in chaotic combinations of what could be shaded on the mesh
(S for potential surf shading, and P for fid map shading)
Now, P toggles whether the shading selected applies to the pot-surf shading, or the fid-map shading,
so only one thing can get shaded at once.
Also, supplied a menu option to control this now.

Fid-map contours will be drawn when pot contours are not drawn.  For no contours to be drawn, you have
to not draw contours, and toggle drawing of fids