Fix crash - compute global min/max on usage from file window
Loading from the command line didn't produce the crash, as it works slightly
differently (order of operations) - GlobalMinMax loops over meshes added
to windows.  Loading from the Files dialog invoked this function before
adding meshes to windows.

So now we make sure to invoke it when we are done loading files.

Fixes #36

It would crash because the range of the scale was based on
min=FLT_MAX and max=FLT_MIN