Implement subseries display and selection
A subseries is identified as globalfid of type 17.

Implement the ability to identify the start of a subseries in the PickWindow
with a vertical line
Implement the ability to only display one subseries at a time
Implement the ability to use shift-right and shift-left to advance
  forward and backward an entire subseries
Implement the abilty to choose subseries to stack on top of each other.
If display-subseries mode is on, the pick window will:
  draw the current subseries only if no subseries are chosen to stack
  draw all stacked subseries subseries on top of each other if there are