matlab tweaks - nested layers, pts/fac names, pts-only is fine
Apparently some matlab files come nested one layer deep - like the saved array name is the first/only entry in the file.
In the case where there is only one entry, and the entry is NOT the name of a pts array (because that is allowed)
Then cascade down the layer and read it normally

Allow pts-only - so it's okay if there is no connectivity.  Just skip the elements sectiono.

Allow pts, point, node, and nodes for points; and fac, face, and faces for faces

Refactor the code a little to de-dup - handle the cell case up front and then treat it like the struct array